Enhanced Result to included previous Results
Results is now more powerful and useful. Use it to
store a history of Results from previously run methods.
ResultCode will now be write protected from parent Result
if a child Result was added that contained a non-ignorable
ResultCode (i.e. Error) so that the top level also shows
that there was an error.